fix: make peer range query robust to existence of 5u8 key prefixes
The open range `scan_peer_key..` we were passing into `db.range()` was
returning all keys that start with `4u8` *or higher*.

Simple fix: give `5u8` as the upper bound of the query.

It's worth noting that the same should apply to the blob range queries,
I'll try to make a commit for those next.